move scheduling enum file to enums directory
