+++ src/enums/scheduling.ts
